requirements.txt
mitmproxy==5.0.1
pymongo==3.10.1
rdflib==4.2.2
requests==2.22.0
python-dateutil
